# api/docs/swagger.yaml
definitions:
  handler.ApproveGuestbookEntryRequest:
    properties:
      approve:
        type: boolean
    type: object
...
  handler.SubmitContactRequest:
    properties:
      email:
        type: string
      honeypot:
        description: Anti-spam field - should be empty
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  handler.SubmitGuestbookEntryRequest:
    properties:
      display_name:
        type: string
      honeypot:
        description: Anti-spam field - should be empty
        type: string
      message:
        type: string
    type: object
  handler.UpdateContactStatusRequest:
    properties:
      status:
        type: string
    type: object
...
  view.ContactSubmissionResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  view.GuestbookEntryResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_approved:
        type: boolean
      message:
        type: string
      user_id:
        type: string
      user_provider:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/contact:
    get:
      description: List contact submissions (admin only, can filter by status)
      parameters:
      - description: Filter by status (received, replied, archived)
        in: query
        name: status
        type: string
      - description: 'Number of submissions per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of submissions to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of submissions
          schema:
            items:
              $ref: '#/definitions/view.ContactSubmissionResponse'
            type: array
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List contact submissions
      tags:
      - Contact
    post:
      consumes:
      - application/json
      description: Submit a contact form message (public, rate limited to 5 per day
        per email)
      parameters:
      - description: Contact form request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Submission received
          schema:
            $ref: '#/definitions/view.ContactSubmissionResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Submit a contact form
      tags:
      - Contact
  /api/contact/{id}:
    patch:
      consumes:
      - application/json
      description: Update the status of a contact submission (admin only)
      parameters:
      - description: Submission ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateContactStatusRequest'
      responses:
        "204":
          description: Status updated
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Submission not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update contact submission status
      tags:
      - Contact
  /api/guestbook:
    get:
      description: List all approved guestbook entries (public)
      parameters:
      - description: 'Number of entries per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of entries to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of approved entries
          schema:
            items:
              $ref: '#/definitions/view.GuestbookEntryResponse'
            type: array
      summary: List approved guestbook entries
      tags:
      - Guestbook
    post:
      consumes:
      - application/json
      description: Submit a new guestbook entry (authenticated users only, requires
        approval)
      parameters:
      - description: Guestbook entry request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitGuestbookEntryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Entry submitted successfully
          schema:
            $ref: '#/definitions/view.GuestbookEntryResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Rate limit exceeded
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a guestbook entry
      tags:
      - Guestbook
  /api/guestbook/{id}:
    delete:
      description: Delete a guestbook entry (admin only)
      parameters:
      - description: Entry ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Entry deleted successfully
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Entry not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a guestbook entry
      tags:
      - Guestbook
  /api/guestbook/{id}/approve:
    post:
      consumes:
      - application/json
      description: Approve or reject a pending guestbook entry (admin only)
      parameters:
      - description: Entry ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Approve/reject request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ApproveGuestbookEntryRequest'
      responses:
        "204":
          description: Entry status updated
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Entry not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve or reject a guestbook entry
      tags:
      - Guestbook
  /api/guestbook/pending:
    get:
      description: List all pending guestbook entries awaiting approval (admin only)
      parameters:
      - description: 'Number of entries per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of entries to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of pending entries
          schema:
            items:
              $ref: '#/definitions/view.GuestbookEntryResponse'
            type: array
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List pending guestbook entries
      tags:
      - Guestbook
...
swagger: "2.0"